Moved data/access_levels/entities to data/account
