+++ src/web/controllers/profile_controller.py
@@ -10,7 +10,7 @@ from litestar.response import Template
 from litestar_htmx import HTMXTemplate
 
 from common.i18n import _
-from data.access_levels.entities import Account
+from data.account import Account
 from data.event import Event
 from database.sqlite.event.event_database import EventDatabase
 from web.controllers.admin.base_admin_controller import AdminWebContext
